+++ include/swift/SIL/ParseTestSpecification.h
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// This file defines Test::Argument.
+// This file defines test::Argument.
 //
 //===----------------------------------------------------------------------===//
