+++ lib/SIL/Parser/CMakeLists.txt
@@ -1,2 +1,3 @@
 target_sources(swiftSIL PRIVATE
+  ParseTestSpecification.cpp
   ParseSIL.cpp)
